+++ packages/expo-modules-core/src/index.ts
@@ -27,7 +27,6 @@ export * from './Refs';
 export * from './hooks/useReleasingSharedObject';
 export * from './hooks/useReleasingSharedObjectWithLifecycle';
 export * from './reload';
-export * from './isRunningOnQuest';
 
 // Errors
 export { CodedError } from './errors/CodedError';